plot_finger.sh, add title and label, file handling, point size
The 2D fingerprint plot now (again) contains d_e and d_i as
in CrystalExplorer.  Gnuplot uses the stem of the input file
again as automatic label on the plot, too.  A now explicit
definition of the point size in the scatter plot, intentionally
lowered in reference to the default setting, prevents these
'tiles' to overlap with each other; so there is less risk to
loose visual information which is present in the .dat file.

Addition of a brief instruction about background and use of
this script.