diff_finger.py, concept study
The number of languages around this project needn't be this
high, probably.  Because I contributed to this situation,
and because there already is the CPython moderator script
hirshfeld_moderator.py present to interact with fingerprint.f90,
diff_finger.c and sum_abs_diffs.rb, this new script probes
how well CPython may address the problem to compute the
difference maps.

Numpy is the sole dependency of this CPython3 script, which is
not met in the standard installation of CPython.