diff_finger.py, consistency check de/di included
The script now determines the coordinates of (x_min,y_min)
of ref_file nd probe_file.  If these are equal till and
including the second decimal .AND. if the line count of the
two files match, then the difference map will be computed.

The access to (x_min, y_min) equally offers automatic re-
insertion of the blank lines in the diff*.dat file written.