update README[.org, .md] about hirshfeld_moderator_windows.py
The script -- working both in Linux, as well in Windows to
assist the computational part of the analysis -- lowering the
external dependencies is mentioned in a footnote.  It no longer
is necessary to have a dedicated C compiler and an installation
of ruby to run the analysis, and given the good work of CPython
and Fortran (e.g., with the free gfortran compiler), there is
little need seen to equally rewrite / translate fingerprint.f90
into fingerprint.py.  Unexpectedly, however, getting Gnuplot.py
to work in Windows represents a larger obstacle, than expected;
mildly suggesting a transition to CPythons matplotlib library
as an alternative.