diff_finger.py, modified re-insertion of blanks
The previous definition to re-insert the blank lines in the same
pattern as provided by diff_finger.c worked well when calling for
CPython3 in Linux, but not when calling for CPython2 in Windows
(which might not work with the current shebang, either).  By the
revised instruction, this issue is fixed.

Thankfully, gnuplot was working fine with the diff*.dat files lacking
these separating lines, though.