diff_finger.py, win_moderator: uniform output format
In previous commits, the column widths about di, de, and area difference reported
was not used consitently, e.g. an entry "de = 2.10" was print as "2.1" only.  To
ease retrieve the entries, the typical record per line for differences equal to
zero or positive is

2.00 2.01 0.00432582

The pattern for entries about negative differences is

2.00 2.02 -0.00439352

with an underlying format of {:3.2f}{:3.2f}{:10.8f}.