hirshfeld_moderator_windows.py, numpy-free diff*.dat generation
The current approach generates the diff_*.dat files by zip-based
simultaneous reading of the file contents.  The script yields
the same results as with the numpy-based approach, with the
advantage of being a bit more compact, to use only modules of
the standard library and hence more portable (including pypy).
Compared to the previous approach, processing is complete in
about half the time; the typical cadence with pypy is close to
one test per second.